requirements.txt
python-multipart>=0.0.6
pillow>=10.2.0,<11.0.0
pydantic-settings==2.7.0
fastapi==0.115.3
uvicorn==0.32.0
gunicorn==23.0.0
pydantic==2.10.4
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.7.0